Add service layer for Canvas with completion functions
